Refactor of the spatiotemporal_cube. Further separation of function by introducing the overarching spatial_engine base class from which 2 children, spatiotemporal_cube and spatiotemporal_lake, originate
