+++ src/bmc/__init__.py
@@ -1,3 +1,4 @@
+from . import engine
 from . import cube
 from . import datasource
 from . import utils
